Added function and ability to draw a proper rectangle
Learned -
1: Everytime we write to the GDDRAM, since this is a 32x128 instead
of the traditional 64x128, it's 4 pixel per page but the page is mapped to
a entire byte. it's the odd digit bits that matter -> 01010101 is 1111
while 11110000 is 1100

2:Serial mode does not support read from display. Everytime we write to display,
it overwrites the content of the 4 pixel unit.

current implementation keeps a memory mapping and their content and directly uses
this memory mapping to write to display

since I'm not sure how the increment of display ram pointer works I manually set the
column and page location each time

Extremely ineffecient but works to draw a rectange for now